create fs_ops
