+++ src/fs.c
@@ -1,4 +1,5 @@
 #include "fs.h"
+#include "fs_ops.h"
 #include "kmalloc.h"
 #include "string.h"
 #include "list.h"
